Use `eth_sign_addr` in sign up validation.
